pyproject.toml
[tool.poetry]
name = "leostaer-bot"
version = "0.1.0"
description = ""
authors = ["TheShepsMT <theshepsmt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
psycopg = "^3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
